+++ flutter/lib/refineui_system_icons.dart
@@ -1,7 +1,7 @@
 library refineui_system_icons;
 
-export 'src/refineui_icons.dart';
 export 'src/refineui_system_icons_base.dart';
+export 'src/refineui_system_icons.dart';
 
 /// RefineUI System Icons for Flutter
 ///