+++ flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.15
+
+- Updated Flutter package with latest RefineUI system icons
+- Fixed icon generation and asset management
+- Improved package structure and documentation
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- TODO: Describe initial release.